Arrange alignment and save BBC ggplot chart

Running this function will save your plot with the correct guidelines for publication for a BBC News graphic. It will left align your title, subtitle and source, add the BBC blocks at the bottom right and save it to your specified location.

finalise_plot(
  plot_name,
  source_name,
  save_filepath = file.path(Sys.getenv("TMPDIR"), "tmp-nc.png"),
  width_pixels = 840,
  height_pixels = 450,
  logo_image_path = system.file("logo.png", package = "FlowmindeR")
)

Arguments

plot_name
The variable name of the plot you have created that you want to format and save

source_name
The text you want to come after the text 'Source:' in the bottom left hand side of your side

save_filepath
Exact filepath that you want the plot to be saved to

width_pixels
Width in pixels that you want to save your chart to - defaults to 640

height_pixels
Height in pixels that you want to save your chart to - defaults to 450

logo_image_path
File path for the logo image you want to use in the right hand side of your chart, which needs to be a PNG file - defaults to BBC blocks image that sits within the data folder of your package

Value

(Invisibly) an updated ggplot object.

Examples

finalise_plot(plot_name = myplot,
source = "The source for my data",
save_filepath = "filename_that_my_plot_should_be_saved_to-nc.png",
width_pixels = 640,
height_pixels = 450,
logo_image_path = "logo_image_filepath.png"
)

#> Error in png::readPNG(logo_image_path): unable to open logo_image_filepath.png